refactor: use-popper hook
