+++ packages/popper/src/index.tsx
@@ -1,2 +1,2 @@
 export * from "./use-popper"
-export { toTransformOrigin } from "./popper.utils"
+export { toTransformOrigin } from "./utils"
